# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/mockRule/addRule": {
            "post": {
                "description": "拦截规则新增时的一些安全检查，不重复检查等",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "新增拦截规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新增的实体",
                        "name": "proxyRulesModel",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mockRule/deleteRulesById": {
            "post": {
                "description": "逻辑删除一个拦截规则",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "删除一个拦截规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "拦截规则的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mockRule/getAllRules": {
            "get": {
                "description": "无差别获取，不根据用户进行隔离",
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "获取当前所有的规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mockRule/getRuleByCons": {
            "get": {
                "description": "根据拦截规则的其中一个或者多个参数来进行筛选，多个参数同时满足时才会返回",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "根据指定的条件来获取一部分拦截规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要满足的多个条件json",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mockRule/getRuleDetailById": {
            "get": {
                "description": "根据拦截规则的其中一个或者多个参数来进行筛选，多个参数同时满足时才会返回",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "根据指定的条件来获取一部分拦截规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "拦截规则的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mockRule/updateRulesById": {
            "post": {
                "description": "更新一个拦截规则的所有信息",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "拦截规则相关接口"
                ],
                "summary": "根据id更新一个拦截规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ownerName",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "拦截规则的id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"message\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}